requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
sqlalchemy>=2.0.0
pyyaml>=6.0
apscheduler>=3.10.0
requests>=2.31.0
